SWEEP: Migrated all exception serialization tests to Lucene.Net.AllProjects and refactored to provide base classes for additional assembly scanning tests
